requirements_dev.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
pytz==2023.3.post1
mysqlclient==2.2.0
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
django-filter==23.2